+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Features:
     the Composer autoloader. #562
   - [WorseReflection] Support trait alias maps. #540
   - [RPC] Return semantic RPC protocol version in response (starting at `1.0.0`).
+  - [Completion] Complete constructor parameters.
 
 Improvements: